+++ messaging-core/src/main/java/com/nabla/sdk/core/data/apollo/ApolloNoDataException.kt
@@ -1,0 +1,3 @@
+package com.nabla.sdk.core.data.apollo
+
+internal class ApolloNoDataException(internal val requestId: String?) : RuntimeException()