+++ messaging-core/src/main/java/com/nabla/sdk/core/data/exception/BaseExceptionMapper.kt
@@ -1,5 +1,6 @@
 package com.nabla.sdk.core.data.exception
 
+import com.nabla.sdk.core.data.apollo.ApolloNoDataException
 import com.nabla.sdk.core.domain.entity.NablaException
 
 internal class BaseExceptionMapper : ExceptionMapper {
@@ -12,6 +13,12 @@ internal class BaseExceptionMapper : ExceptionMapper {
         return when {
             exception is NablaException -> exception
             exception.isNetworkError() -> NablaException.Network(exception)
+            exception is ApolloNoDataException -> NablaException.Server(
+                exception,
+                code = -1,
+                serverMessage = "The server did not return any data",
+                requestId = exception.requestId,
+            )
             exception is GraphQLException -> NablaException.Server(
                 exception,
                 exception.numericCode ?: 0,